requirements.txt
asgiref==3.11.0
django==5.2
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
pytz==2025.2
pyyaml==6.0.3
sqlparse==0.5.4
uritemplate==4.2.0
